requirements.txt
mysql-connector-python==8.2.0
nltk==3.8.1
PySastrawi==1.2.0
scikit_learn==1.3.2
shortuuid==1.0.11
fastapi[all]
